+++ demo/js/nguibs.js
@@ -34,6 +34,24 @@
 	}]);
 
 	angular.module('demoApp')
+	.controller('paginationCtrl', ['$scope', function($scope){
+		$scope.totalItems = 64;
+		$scope.currentPage = 4;
+
+		$scope.setPage = function (pageNo) {
+		  $scope.currentPage = pageNo;
+		};
+
+		$scope.pageChanged = function() {
+		  $log.log('Page changed to: ' + $scope.currentPage);
+		};
+
+		$scope.maxSize = 5;
+		$scope.bigTotalItems = 175;
+		$scope.bigCurrentPage = 1;
+	}]);
+
+	angular.module('demoApp')
 	.controller('datepickerCtrl', ['$scope', function($scope){
 	}]);